fix(gameController): change how to retrieve player
Now the gameController provides 2 methods for
retrieving players. The first one is the method
used to retrieve current player, the second one
is the method used to retrieve their opponent.

Besides, I add a method called playWithComputer()
to the gameController. With this method, a new
computerized player will be added to the player
list and becomes the opponent the current player.