feat(player): add computerMakeChoice() method
This method will depend on the opponent's
gameboard grid and help the computer to compute
the next attack coordinates in order to make it
possible to hit player's ship.

Besides, I add a new method to this class called
clearBoard(), this method will create a new
gameboard for the player when being called.