fix(gameboard): change how to managing shots
Previously, shots are stored in a separated array
from the grid. Now I also add the shots to the
position on the grid that they are fired on by
players. This help optimizing performance when
rendering grid and shots into view since if only
storing shots in the separated array, it is
neccessary to loop through that array after
looping through the grid itself.