feat(gameplay): add sunk ship notification
When a player sinks the others and have ship sunk,
a short message will appear to inform them.

Now receiveAttack() will return the result of that
attack (of type hit or miss) as well as the ship
that the attack has sunk (if there is).

I also change the message when the player receive
attacks to be uniform with the message when the
player make attacks.